Move horrible output helper methods
